src/requirements.txt
beautifulsoup4==4.14.3
requests==2.33.1
google-generativeai==0.8.6
python-dotenv==1.2.2
pymysql==1.1.0
playwright==1.42.0
httpx==0.27.0
openai==1.14.0
